start with a good refactor
* split up into multiple files
* split up data fetching and manipulation into multiple effects
* adapt for new subgraph
