+++ src/components/Footer/MarketStats/_queries.ts
@@ -1,0 +1,20 @@
+import { gql } from 'urql'
+
+export const queryGlobalStatistics = gql`
+  query FooterStatsValues {
+    globalStatistics {
+      poolCount
+      nftCount
+      datatokenCount
+      orderCount
+      totalLiquidity {
+        value
+        token {
+          address
+          name
+          symbol
+        }
+      }
+    }
+  }
+`
