+++ src/components/Footer/MarketStats/index.module.css
@@ -1,0 +1,16 @@
+.stats {
+  margin-bottom: calc(var(--spacer) * 2);
+}
+
+/* specificity sledgehammer override without !important */
+.stats,
+.stats *,
+.statsList * {
+  font-size: var(--font-size-small);
+  color: var(--color-secondary);
+  margin-left: 0;
+}
+
+.info {
+  width: 0.85rem;
+}
